requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==3.2.12
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
